requirements.txt
Flask==3.0.3
tensorflow==2.19.0
keras==3.10.0
numpy
Pillow
